requirements.txt
vertica-python
pyyaml

